cool new alex UI now with experience omg good job alex the UI is great!!

+++ ShrinesOfElementia/Assets/Scripts/EventSystem/Events/ExperienceEvent.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class ExperienceEvent : DebugEvent
 {
-    public int Experience;
+    public float Experience;
 
-    public ExperienceEvent(string eventDescription, int experience) : base(eventDescription)
+    public ExperienceEvent(string eventDescription, float experience) : base(eventDescription)
     {
         Experience = experience;
     }
